Time: 22 ms (38.63%), Space: 21.6 MB (88.84%) - LeetHub

+++ 98-validate-binary-search-tree/98-validate-binary-search-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    bool inRange(TreeNode *root , long long minV , long long maxV){
+    bool inRange(TreeNode *root , long  minV , long  maxV){
     if(!root){return true;}
     if(root->val <= minV || root->val >= maxV){
         return false;
